1662 Fix Exim panics

+++ installer/fix_exim4_panic.sh
@@ -1,6 +1,5 @@
 #!/bin/sh
 
-sudo vim /etc/exim4/update-exim4.conf.conf
-# dc_local_interfaces='127.0.0.1 ; ::1' -> dc_local_interfaces='127.0.0.1'
+sudo sed -i "s/dc_local_interfaces='127.0.0.1 ; ::1'/dc_local_interfaces='127.0.0.1'/" /etc/exim4/update-exim4.conf.conf
 
-sudo rm /var/log/exim4/paniclog
+test -f /var/log/exim4/paniclog && sudo rm /var/log/exim4/paniclog
